# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BeatSight Backend",
    "description": "\n## BeatSight Web API\n\nBeatSight is an AI-powered drum beatmap generation platform. This API provides:\n\n### Features\n- **Authentication**: JWT-based auth with access/refresh tokens\n- **Songs**: Upload, manage, and organize your music library\n- **AI Jobs**: Queue songs for AI beatmap generation with real-time progress\n- **Cloud Sync**: Sync preferences and progress across devices\n- **Storage**: Secure file upload with presigned URLs\n\n### Authentication\nMost endpoints require authentication. Include the JWT token in the Authorization header:\n```\nAuthorization: Bearer <access_token>\n```\n\n### Rate Limits\n- Standard users: 100 requests/minute\n- Premium users: 500 requests/minute\n- AI job quota: Based on subscription tier\n    ",
    "contact": {
      "name": "BeatSight Support",
      "url": "https://beatsight.app/support",
      "email": "support@beatsight.app"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/karma/me/roles/{role_code}": {
      "post": {
        "tags": [
          "karma"
        ],
        "summary": "Claim Role",
        "description": "Claim an eligible role.\n\nUser must meet the karma threshold and any verification requirements.",
        "operationId": "claim_role_api_karma_me_roles__role_code__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "role_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Role Code"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Claim Role Api Karma Me Roles  Role Code  Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/users/{user_id}/role": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Update user role",
        "description": "Add or remove a role from a user.",
        "operationId": "update_user_role_api_admin_users__user_id__role_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRoleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/roles": {
      "get": {
        "tags": [
          "roles"
        ],
        "summary": "List Roles",
        "description": "List all available roles.\n\nAny authenticated user can see the list of roles.",
        "operationId": "list_roles_api_roles_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RoleInfo"
                  },
                  "type": "array",
                  "title": "Response List Roles Api Roles Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/roles/my-roles": {
      "get": {
        "tags": [
          "roles"
        ],
        "summary": "Get My Roles",
        "description": "Get the current user's roles and permissions.",
        "operationId": "get_my_roles_api_roles_my_roles_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRolesResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/roles/users/{user_id}": {
      "get": {
        "tags": [
          "roles"
        ],
        "summary": "Get User Roles",
        "description": "Get a specific user's roles and permissions.\n\nRequires ROLE_LIST permission (admin only).",
        "operationId": "get_user_roles_api_roles_users__user_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRolesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/roles/assign": {
      "post": {
        "tags": [
          "roles"
        ],
        "summary": "Assign Role",
        "description": "Assign a role to a user.\n\nRequires ROLE_ASSIGN permission (admin only).\n\nAdmins cannot assign the admin role to themselves for safety.",
        "operationId": "assign_role_api_roles_assign_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleAssignRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/roles/revoke": {
      "post": {
        "tags": [
          "roles"
        ],
        "summary": "Revoke Role",
        "description": "Revoke a role from a user.\n\nRequires ROLE_REVOKE permission (admin only).\n\nAdmins cannot revoke their own admin role for safety.\nCannot revoke the 'user' role as it's the default.",
        "operationId": "revoke_role_api_roles_revoke_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleRevokeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/roles/permissions": {
      "get": {
        "tags": [
          "roles"
        ],
        "summary": "List Permissions",
        "description": "List all available permissions in the system.\n\nAny authenticated user can see the list of permissions.",
        "operationId": "list_permissions_api_roles_permissions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array",
                  "title": "Response List Permissions Api Roles Permissions Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/roles/check/{permission}": {
      "get": {
        "tags": [
          "roles"
        ],
        "summary": "Check Permission",
        "description": "Check if the current user has a specific permission.\n\nUseful for frontend to determine UI state.",
        "operationId": "check_permission_api_roles_check__permission__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "permission",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Permission"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Check Permission Api Roles Check  Permission  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminUserActionResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "title": "AdminUserActionResponse",
        "description": "Response for user management actions."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RoleActionResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "role_code": {
            "type": "string",
            "title": "Role Code"
          }
        },
        "type": "object",
        "required": [
          "success",
          "message",
          "user_id",
          "role_code"
        ],
        "title": "RoleActionResponse",
        "description": "Response for role assignment/revocation."
      },
      "RoleAssignRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "role_code": {
            "type": "string",
            "title": "Role Code"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "role_code"
        ],
        "title": "RoleAssignRequest",
        "description": "Request to assign a role to a user."
      },
      "RoleInfo": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "min_karma": {
            "type": "integer",
            "title": "Min Karma"
          },
          "requires_phone_verification": {
            "type": "boolean",
            "title": "Requires Phone Verification"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Permissions"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "description",
          "min_karma",
          "requires_phone_verification",
          "created_at",
          "permissions"
        ],
        "title": "RoleInfo",
        "description": "Information about a role."
      },
      "RoleRevokeRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "role_code": {
            "type": "string",
            "title": "Role Code"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "role_code"
        ],
        "title": "RoleRevokeRequest",
        "description": "Request to revoke a role from a user."
      },
...
      "UpdateUserRoleRequest": {
        "properties": {
          "role": {
            "type": "string",
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "role"
        ],
        "title": "UpdateUserRoleRequest",
        "description": "Request to update user role."
      },
...
      "UserRoleInfo": {
        "properties": {
          "role_code": {
            "type": "string",
            "title": "Role Code"
          },
          "assigned_at": {
            "type": "string",
            "format": "date-time",
            "title": "Assigned At"
          }
        },
        "type": "object",
        "required": [
          "role_code",
          "assigned_at"
        ],
        "title": "UserRoleInfo",
        "description": "User's role assignment info."
      },
      "UserRolesResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "roles": {
            "items": {
              "$ref": "#/components/schemas/UserRoleInfo"
            },
            "type": "array",
            "title": "Roles"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Permissions"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "roles",
          "permissions"
        ],
        "title": "UserRolesResponse",
        "description": "Response with user's roles and permissions."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}